feat: extract disclosure
