+++ src/connect/components/disclosure/disclosure.stories.tsx
@@ -1,0 +1,40 @@
+import { useArgs } from '@storybook/preview-api';
+import { Meta, StoryObj } from '@storybook/react';
+import { Disclosure } from '.';
+
+const meta = {
+    title: 'Connect/Components/Disclosure',
+    component: Disclosure,
+} satisfies Meta<typeof Disclosure>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+    args: {
+        title: 'Disclosure',
+        children: (
+            <p>
+                lorem ipsum dolor sit amet consectetur adipisicing elit.
+                Quisquam
+            </p>
+        ),
+        isOpen: false,
+        onOpenChange: () => {},
+    },
+    render: function Wrapper(args) {
+        const [, setArgs] = useArgs<typeof args>();
+
+        return (
+            <div className="w-[408px] bg-white p-6">
+                <Disclosure
+                    {...args}
+                    onOpenChange={() =>
+                        setArgs({ ...args, isOpen: !args.isOpen })
+                    }
+                />
+            </div>
+        );
+    },
+};
