+++ src/connect/components/disclosure/index.ts
@@ -1,0 +1,1 @@
+export * from './disclosure';
