--- src/connect/components/drive-settings-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './drive-settings-modal';
