+++ src/connect/components/form/drive-settings-form.tsx
@@ -2,10 +2,10 @@ import { Divider, DriveLocation, SharingType } from '@/connect';
 import { Button, Icon } from '@/powerhouse';
 import { useState } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
+import { Disclosure } from '../disclosure';
 import { DriveNameInput } from '../drive-name-input/drive-name-input';
 import { AvailableOfflineToggle } from './available-offline-toggle';
 import { DeleteDrive } from './delete-drive';
-import { Disclosure } from './disclosure';
 import { Label } from './label';
 import { LocationInfo } from './location-info';
 import { SharingTypeFormInput } from './sharing-type-form-input';
