+++ src/connect/components/modal/index.ts
@@ -1,2 +1,5 @@
+export * from './add-public-drive-modal';
+export * from './create-local-drive-modal';
 export * from './delete-item-modal';
+export * from './drive-settings-modal';
 export * from './upgrade-drive-modal';
